Fixed correlations dtypes issue

+++ src/dashboard/pages/page_2.py
@@ -87,14 +87,14 @@ def load_correlations(start_time, end_time):
     autocorr_sign = np.zeros(k_max)
     autocorr_size = np.zeros(k_max)
     counts = np.zeros(k_max)
-    signs = np.empty(CHUNK_SIZE+k_max, dtype=np.int8)
+    signs = np.empty(CHUNK_SIZE+k_max, dtype=np.float32)
     sizes = np.empty(CHUNK_SIZE+k_max, dtype=np.float32)
-    signs[:k_max] *= 0
+    signs[:k_max] *= 0.0
     sizes[:k_max] *= 0.0
     start = 0
     for chunk in read_trades_in_chunks(start_time, end_time):
         n = len(chunk)
-        signs[k_max:n+k_max] = chunk["sign"].to_numpy(dtype=np.int8, copy=False)
+        signs[k_max:n+k_max] = chunk["sign"].to_numpy(dtype=np.float32, copy=False)
         sizes[k_max:n+k_max] = signs[k_max:n+k_max] * chunk["quantity"].to_numpy(dtype=np.float32, copy=False)
         for i in range(1, k_max+1):
             autocorr_sign[i-1] += np.dot(signs[i:n+k_max], signs[:n+k_max-i])
